# SIMD: Get rid of attribute-based CPU dispatching