--- a/numpy/core/code_generators/generate_umath.py
+++ b/numpy/core/code_generators/generate_umath.py
@@ -61,9 +61,6 @@ class TypeDescription:
     cfunc_alias : str or none, optional
         Appended to inner loop C function name, e.g., FLOAT_{cfunc_alias}. See make_arrays.
         NOTE: it doesn't support 'astype'
-    simd : list
-        Available SIMD ufunc loops, dispatched at runtime in specified order
-        Currently only supported for simples types (see make_arrays)
     dispatch : str or None, optional
         Dispatch-able source name without its extension '.dispatch.c' that
         contains the definition of ufunc, dispatched at runtime depending on the
@@ -71,7 +68,7 @@ class TypeDescription:
         NOTE: it doesn't support 'astype'
     """
     def __init__(self, type, f=None, in_=None, out=None, astype=None, cfunc_alias=None,
-                 simd=None, dispatch=None):
+                 dispatch=None):
         self.type = type
         self.func_data = f
         if astype is None:
@@ -84,7 +81,6 @@ class TypeDescription:
             out = out.replace('P', type)
         self.out = out
         self.cfunc_alias = cfunc_alias
-        self.simd = simd
         self.dispatch = dispatch
 
     def finish_signature(self, nin, nout):
@@ -146,8 +142,9 @@ def build_func_data(types, f):
     return func_data
 
 def TD(types, f=None, astype=None, in_=None, out=None, cfunc_alias=None,
-       simd=None, dispatch=None):
-    """Generate a TypeDescription instance for each item in types
+       dispatch=None):
+    """
+    Generate a TypeDescription instance for each item in types
     """
     if f is not None:
         if isinstance(f, str):
@@ -172,12 +169,6 @@ def TD(types, f=None, astype=None, in_=None, out=None, cfunc_alias=None,
         raise ValueError("Number of types and outputs do not match")
     tds = []
     for t, fd, i, o in zip(types, func_data, in_, out):
-        # [(simd-name, list of types)]
-        if simd is not None:
-            simdt = [k for k, v in simd if t in v]
-        else:
-            simdt = []
-
         # [(dispatch file name without extension '.dispatch.c*', list of types)]
         if dispatch:
             dispt = ([k for k, v in dispatch if t in v]+[None])[0]
@@ -185,7 +176,7 @@ def TD(types, f=None, astype=None, in_=None, out=None, cfunc_alias=None,
             dispt = None
         tds.append(TypeDescription(
             t, f=fd, in_=i, out=o, astype=astype, cfunc_alias=cfunc_alias,
-            simd=simdt, dispatch=dispt
+            dispatch=dispt
         ))
     return tds
 
@@ -352,8 +343,10 @@ defdict = {
           docstrings.get('numpy.core.umath.add'),
           'PyUFunc_AdditionTypeResolver',
           TD('?', cfunc_alias='logical_or', dispatch=[('loops_logical', '?')]),
-          TD(no_bool_times_obj, simd=[('avx2', ints)],
-                                dispatch=[('loops_arithm_fp', 'fdFD')]),
+          TD(no_bool_times_obj, dispatch=[
+              ('loops_arithm_fp', 'fdFD'),
+              ('loops_autovec_int', ints),
+          ]),
           [TypeDescription('M', FullTypeDescr, 'Mm', 'M'),
            TypeDescription('m', FullTypeDescr, 'mm', 'm'),
            TypeDescription('M', FullTypeDescr, 'mM', 'M'),
@@ -365,8 +358,10 @@ defdict = {
     Ufunc(2, 1, None, # Zero is only a unit to the right, not the left
           docstrings.get('numpy.core.umath.subtract'),
           'PyUFunc_SubtractionTypeResolver',
-          TD(no_bool_times_obj, simd=[('avx2', ints)],
-                                dispatch=[('loops_arithm_fp', 'fdFD')]),
+          TD(no_bool_times_obj, dispatch=[
+              ('loops_arithm_fp', 'fdFD'),
+              ('loops_autovec_int', ints),
+          ]),
           [TypeDescription('M', FullTypeDescr, 'Mm', 'M'),
            TypeDescription('m', FullTypeDescr, 'mm', 'm'),
            TypeDescription('M', FullTypeDescr, 'MM', 'm'),
@@ -380,8 +375,10 @@ defdict = {
           'PyUFunc_MultiplicationTypeResolver',
           TD('?', cfunc_alias='logical_and',
                   dispatch=[('loops_logical', '?')]),
-          TD(no_bool_times_obj, simd=[('avx2', ints)],
-                                dispatch=[('loops_arithm_fp', 'fdFD')]),
+          TD(no_bool_times_obj, dispatch=[
+              ('loops_arithm_fp', 'fdFD'),
+              ('loops_autovec_int', ints),
+          ]),
           [TypeDescription('m', FullTypeDescr, 'mq', 'm'),
            TypeDescription('m', FullTypeDescr, 'qm', 'm'),
            TypeDescription('m', FullTypeDescr, 'md', 'm'),
@@ -421,8 +418,10 @@ defdict = {
     Ufunc(1, 1, None,
           docstrings.get('numpy.core.umath.conjugate'),
           None,
-          TD(ints+flts+cmplx, simd=[('avx2', ints)],
-            dispatch=[('loops_arithm_fp', 'FD')]),
+          TD(ints+flts+cmplx, dispatch=[
+              ('loops_arithm_fp', 'FD'),
+              ('loops_autovec_int', ints),
+          ]),
           TD(P, f='conjugate'),
           ),
 'fmod':
@@ -437,15 +436,21 @@ defdict = {
     Ufunc(1, 1, None,
           docstrings.get('numpy.core.umath.square'),
           None,
-          TD(ints+inexact, simd=[('avx2', ints)],
-             dispatch=[('loops_unary_fp', 'fd'), ('loops_arithm_fp', 'FD')]),
+          TD(ints+inexact, dispatch=[
+              ('loops_unary_fp', 'fd'),
+              ('loops_arithm_fp', 'FD'),
+              ('loops_autovec_int', ints),
+          ]),
           TD(O, f='Py_square'),
           ),
 'reciprocal':
     Ufunc(1, 1, None,
           docstrings.get('numpy.core.umath.reciprocal'),
           None,
-          TD(ints+inexact, simd=[('avx2', ints)], dispatch=[('loops_unary_fp', 'fd')]),
+          TD(ints+inexact, dispatch=[
+              ('loops_unary_fp', 'fd'),
+              ('loops_autovec_int', ints),
+          ]),
           TD(O, f='Py_reciprocal'),
           ),
 # This is no longer used as numpy.ones_like, however it is
@@ -563,24 +568,30 @@ defdict = {
     Ufunc(2, 1, True_,
           docstrings.get('numpy.core.umath.logical_and'),
           'PyUFunc_SimpleBinaryComparisonTypeResolver',
-          TD(nodatetime_or_obj, out='?', simd=[('avx2', ints)],
-                                dispatch=[('loops_logical', '?')]),
+          TD(nodatetime_or_obj, out='?', dispatch=[
+              ('loops_logical', '?'),
+              ('loops_autovec_int', ints),
+          ]),
           TD(O, f='npy_ObjectLogicalAnd'),
           ),
 'logical_not':
     Ufunc(1, 1, None,
           docstrings.get('numpy.core.umath.logical_not'),
           None,
-          TD(nodatetime_or_obj, out='?', simd=[('avx2', ints)],
-                                dispatch=[('loops_logical', '?')]),
+          TD(nodatetime_or_obj, out='?', dispatch=[
+              ('loops_logical', '?'),
+              ('loops_autovec_int', ints),
+          ]),
           TD(O, f='npy_ObjectLogicalNot'),
           ),
 'logical_or':
     Ufunc(2, 1, False_,
           docstrings.get('numpy.core.umath.logical_or'),
           'PyUFunc_SimpleBinaryComparisonTypeResolver',
-          TD(nodatetime_or_obj, out='?', simd=[('avx2', ints)],
-                                dispatch=[('loops_logical', '?')]),
+          TD(nodatetime_or_obj, out='?', dispatch=[
+              ('loops_logical', '?'),
+              ('loops_autovec_int', ints),
+          ]),
           TD(O, f='npy_ObjectLogicalOr'),
           ),
 'logical_xor':
@@ -656,7 +667,7 @@ defdict = {
           None,
           TD('?', cfunc_alias='logical_and',
                   dispatch=[('loops_logical', '?')]),
-          TD(ints, simd=[('avx2', ints)]),
+          TD(ints, dispatch=[('loops_autovec_int', ints)]),
           TD(O, f='PyNumber_And'),
           ),
 'bitwise_or':
@@ -664,7 +675,7 @@ defdict = {
           docstrings.get('numpy.core.umath.bitwise_or'),
           None,
           TD('?', cfunc_alias='logical_or', dispatch=[('loops_logical', '?')]),
-          TD(ints, simd=[('avx2', ints)]),
+          TD(ints, dispatch=[('loops_autovec_int', ints)]),
           TD(O, f='PyNumber_Or'),
           ),
 'bitwise_xor':
@@ -673,7 +684,7 @@ defdict = {
           None,
           TD('?', cfunc_alias='not_equal',
                   dispatch=[('loops_comparison', '?')]),
-          TD(ints, simd=[('avx2', ints)]),
+          TD(ints, dispatch=[('loops_autovec_int', ints)]),
           TD(O, f='PyNumber_Xor'),
           ),
 'invert':
@@ -682,21 +693,21 @@ defdict = {
           None,
           TD('?', cfunc_alias='logical_not',
                   dispatch=[('loops_logical', '?')]),
-          TD(ints, simd=[('avx2', ints)]),
+          TD(ints, dispatch=[('loops_autovec_int', ints)]),
           TD(O, f='PyNumber_Invert'),
           ),
 'left_shift':
     Ufunc(2, 1, None,
           docstrings.get('numpy.core.umath.left_shift'),
           None,
-          TD(ints, simd=[('avx2', ints)]),
+          TD(ints, dispatch=[('loops_autovec_int', ints)]),
           TD(O, f='PyNumber_Lshift'),
           ),
 'right_shift':
     Ufunc(2, 1, None,
           docstrings.get('numpy.core.umath.right_shift'),
           None,
-          TD(ints, simd=[('avx2', ints)]),
+          TD(ints, dispatch=[('loops_autovec_int', ints)]),
           TD(O, f='PyNumber_Rshift'),
           ),
 'heaviside':
@@ -1156,18 +1167,6 @@ def make_arrays(funcdict):
                 datalist.append('(void *)NULL')
                 tname = english_upper(chartoname[t.type])
                 cfunc_fname = f"{tname}_{cfunc_alias}"
-                if t.simd is not None:
-                    for vt in t.simd:
-                        code2list.append(textwrap.dedent("""\
-                        #ifdef HAVE_ATTRIBUTE_TARGET_{ISA}
-                        if (NPY_CPU_HAVE({ISA})) {{
-                            {fname}_functions[{idx}] = {cname}_{isa};
-                        }}
-                        #endif
-                        """).format(
-                            ISA=vt.upper(), isa=vt,
-                            fname=name, cname=cfunc_fname, idx=k
-                        ))
             else:
                 try:
                     thedict = arity_lookup[uf.nin, uf.nout]