… benchmarks/benchmarks/bench_ufunc_strides.py
… benchmarks/benchmarks/common.py
… numpy/core/config.h.in
… numpy/core/setup_common.py
… numpy/core/src/common/simd/avx2/memory.h
… numpy/core/src/common/simd/sse/memory.h
… numpy/core/src/common/simd/vec/conversion.h
… numpy/core/src/common/simd/vec/math.h
… numpy/core/src/common/simd/vec/memory.h
… numpy/core/src/common/simd/vec/misc.h
… numpy/core/src/npysort/mergesort.cpp
… numpy/core/src/umath/fast_loop_macros.h
… numpy/core/src/umath/loops.c.src
… numpy/core/src/umath/loops.h.src
… numpy/core/src/umath/loops_autovec_int.dispatch.c.src