@mattip 2023-02-06
Looks like a nice code cleanup. Do benchmarks confirm that this does not change anything?