@seiko2plus 2023-02-22
> Looks like a nice code cleanup. Do benchmarks confirm that this does not change anything?

I had to enhance the benchmark code to answer this question. No regression, please check the pr description, I have updated it.